wip: move common types to root
