+++ src/kairos/workspace/application/insert-workspace.type.ts
@@ -1,6 +1,6 @@
 import { Result } from '@/kairos/shared/result/result.type';
-import type { WorkspaceSlugConflictError } from '@/kairos/workspace/domain/errors';
-import type { Workspace } from '@/kairos/workspace/domain/workspace.type';
+import type { WorkspaceSlugConflictError } from '@/kairos/workspace/errors';
+import type { Workspace } from '@/kairos/workspace/workspace.type';
 
 export interface WorkspaceToInsert {
   environments: string[];
