+++ src/kairos/workspace/errors.ts
