+++ src/kairos/workspace/infrastructure/repository/insert-workspace.test.ts
@@ -1,6 +1,6 @@
 import { workspacesCollection } from '@/framework/mongodb/collection/workspaces-collection';
 import type { WorkspacesCollection } from '@/framework/mongodb/schema/workspaces-collection-schema';
-import { workspaceSlugConflictError } from '@/kairos/workspace/domain/errors';
+import { workspaceSlugConflictError } from '@/kairos/workspace/errors';
 import { makeInsertWorkspace } from '@/kairos/workspace/infrastructure/repository/insert-workspace';
 import { expectAsyncToThrow } from '@tests/__vitest__/expect-async-to-throw';
 import { integrationTest } from '@tests/__vitest__/integration-test';
