+++ src/kairos/workspace/infrastructure/repository/insert-workspace.ts
@@ -3,7 +3,7 @@ import { WorkspacesCollection } from '@/framework/mongodb/schema/workspaces-coll
 import { err } from '@/kairos/shared/result/err';
 import { ok } from '@/kairos/shared/result/ok';
 import { type InsertWorkspace } from '@/kairos/workspace/application/insert-workspace.type';
-import { workspaceSlugConflictError } from '@/kairos/workspace/domain/errors';
+import { workspaceSlugConflictError } from '@/kairos/workspace/errors';
 
 export const makeInsertWorkspace =
   (collection: WorkspacesCollection): InsertWorkspace =>
